insertion d'un client et insertion d'une vente

+++ dvdstore/src/main/java/com/simplon/dvdstore/dvdstore/repositories/DvdStoreRepositoryModelSales.java
@@ -19,5 +19,5 @@ public class DvdStoreRepositoryModelSales {
     @Column(name="fk_movies")
     private int FKMovies;
     @Column(name="quantity_sales")
-    private float quantityOfSales;
+    private int quantityOfSales;
 }
